Adding routines to dump tables to csv for quick inspect
